feat(laboratories): make laboratories editable
